backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "bcrypt>=4.0.0,<5.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]

[tool.pyright]
pythonVersion = "3.14"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 88
